requirements.txt
Flask==0.12.2
gunicorn==19.7.1
requests==2.17.3
urllib3==1.21.1
virtualenv==15.1.0
smartfile==2.19
psycopg2==2.7.1
beautifulsoup4==4.6.0
python-dateutil==2.6.1
pycrypto==2.6.1
numpy==1.13.1